Remove results.json, move types to types.ts, remove results validation from benchmark
