--- packages/solver/benchmarks/results/results.json
@@ -1,128 +1,0 @@
-[
-  {
-    "blanksCount": 0,
-    "durations": [
-      111.02,
-      95.88,
-      93.85,
-      95.03,
-      95.15
-    ],
-    "label": "English (US)",
-    "locale": "en-US",
-    "resultsCount": 1559,
-    "medianDuration": 95.15
-  },
-  {
-    "blanksCount": 1,
-    "durations": [
-      1037.69,
-      1053.9,
-      1048.83,
-      1055.25,
-      1053
-    ],
-    "label": "English (US)",
-    "locale": "en-US",
-    "resultsCount": 7546,
-    "medianDuration": 1053
-  },
-  {
-    "blanksCount": 2,
-    "durations": [
-      7885.54,
-      7721.24,
-      7795.16,
-      7874.95,
-      7971.14
-    ],
-    "label": "English (US)",
-    "locale": "en-US",
-    "resultsCount": 25659,
-    "medianDuration": 7874.95
-  },
-  {
-    "blanksCount": 0,
-    "durations": [
-      113.15,
-      113.24,
-      111.89,
-      115.95,
-      114.48
-    ],
-    "label": "English (GB)",
-    "locale": "en-GB",
-    "resultsCount": 2398,
-    "medianDuration": 113.24
-  },
-  {
-    "blanksCount": 1,
-    "durations": [
-      1147.77,
-      1136.76,
-      1153.12,
-      1160.56,
-      1184.6
-    ],
-    "label": "English (GB)",
-    "locale": "en-GB",
-    "resultsCount": 11680,
-    "medianDuration": 1153.12
-  },
-  {
-    "blanksCount": 2,
-    "durations": [
-      9833.32,
-      10120.49,
-      10176.96,
-      9860.29,
-      10140.84
-    ],
-    "label": "English (GB)",
-    "locale": "en-GB",
-    "resultsCount": 39286,
-    "medianDuration": 10120.49
-  },
-  {
-    "blanksCount": 0,
-    "durations": [
-      96.99,
-      96.66,
-      85.84,
-      73.53,
-      72.87
-    ],
-    "label": "Polish",
-    "locale": "pl-PL",
-    "resultsCount": 1826,
-    "medianDuration": 85.84
-  },
-  {
-    "blanksCount": 1,
-    "durations": [
-      1146.72,
-      1161.99,
-      1154.14,
-      1137.78,
-      1137.18
-    ],
-    "label": "Polish",
-    "locale": "pl-PL",
-    "resultsCount": 15289,
-    "medianDuration": 1146.72
-  },
-  {
-    "blanksCount": 2,
-    "durations": [
-      13135.85,
-      13296.09,
-      13275.24,
-      13003.19,
-      13475.83
-    ],
-    "label": "Polish",
-    "locale": "pl-PL",
-    "resultsCount": 70691,
-    "medianDuration": 13275.24
-  }
-]
