+++ packages/solver/benchmarks/scenarios.ts
@@ -1,12 +1,6 @@
 import { Locale } from '@scrabble-solver/types';
 
-export interface LanguageScenario {
-  baseRack: string[];
-  boardRows: string[];
-  boardWords: string[];
-  label: string;
-  locale: Locale;
-}
+import { type LanguageScenario } from './types';
 
 const ENGLISH_BOARD_ROWS = [
   '               ',
@@ -25,8 +19,6 @@ const ENGLISH_BOARD_ROWS = [
   '    n          ',
   '               ',
 ];
-
-const ENGLISH_BOARD_WORDS = ['axe', 'cozy', 'flames', 'oaf', 'ratio', 'raven', 'stone', 'taxi', 'verge'];
 
 const ENGLISH_BASE_RACK = ['r', 'e', 't', 'i', 'n', 'a', 's'];
 
@@ -48,29 +40,24 @@ const POLISH_BOARD_ROWS = [
   '               ',
 ];
 
-const POLISH_BOARD_WORDS = ['etos', 'kier', 'korale', 'mysz', 'rak', 'smak', 'tor', 'war', 'wino'];
-
 const POLISH_BASE_RACK = ['a', 'e', 'i', 'n', 'o', 'r', 's'];
 
 export const LANGUAGE_SCENARIOS: LanguageScenario[] = [
   {
     baseRack: ENGLISH_BASE_RACK,
     boardRows: ENGLISH_BOARD_ROWS,
-    boardWords: ENGLISH_BOARD_WORDS,
     label: 'English (US)',
     locale: Locale.EN_US,
   },
   {
     baseRack: ENGLISH_BASE_RACK,
     boardRows: ENGLISH_BOARD_ROWS,
-    boardWords: ENGLISH_BOARD_WORDS,
     label: 'English (GB)',
     locale: Locale.EN_GB,
   },
   {
     baseRack: POLISH_BASE_RACK,
     boardRows: POLISH_BOARD_ROWS,
-    boardWords: POLISH_BOARD_WORDS,
     label: 'Polish',
     locale: Locale.PL_PL,
   },
